Tritium/memory.h: cast void* to intptr_t instead of int
Casting void* to int works fine on 32-bit systems, but typically fails
on a 64-bit stytem.  intptr_t is defined in stdint.h for this very
purpose.

+++ src/Tritium/Tritium/memory.hpp
@@ -24,6 +24,7 @@
 #include <memory>
 #include <boost/shared_ptr.hpp>
 #include <boost/weak_ptr.hpp>
+#include <stdint.h>
 
 /**
  * \brief Header file with smart pointers, etc.
@@ -48,7 +49,7 @@ namespace Tritium
     }
 
     inline bool not_aligned_N(void *ptr, int N) {
-	return (((int)ptr) % N);
+	return (((intptr_t)ptr) % N);
     }
 
     inline bool is_aligned_N(void *ptr, int N) {
